+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * add ```check-consistent-dev-env-version.sh``` to CircleCI pipeline
+* add ```run-bandit.sh``` to CircleCI pipeline
 
 ### Changed